backend/requirements.txt
Flask
Flask-CORS
gunicorn
python-dotenv
google-adk
google-cloud-aiplatform 
googletrans==4.0.0-rc1
